requirements.txt
absl-py==2.1.0
accelerate==1.4.0
aiobotocore==2.5.4
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioitertools==0.12.0
aiosignal==1.3.1
anndata==0.10.9
annotated-types==0.7.0
array_api_compat==1.9.1
asciitree==0.3.3
async-timeout==4.0.3
attrs==24.2.0
backcall==0.2.0
beartype==0.14.1
better-abc==0.0.3
botocore==1.31.17
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
colorcet==3.1.0
contourpy==1.3.0
cycler==0.12.1
dask==2024.10.0
dask-expr==1.1.16
dask-image==2024.5.3
datasets==3.3.2
datashader==0.16.3
dill==0.3.8
distributed==2024.10.0
docker-pycreds==0.4.0
docrep==0.3.2
ecos==2.0.14
einops==0.8.0
executing==1.2.0
faiss==1.9.0
fancy-einsum==0.0.3
fasteners==0.19
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2023.6.0
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
grpcio==1.67.1
h5py==3.12.1
huggingface-hub==0.26.2
idna==3.10
igraph==0.11.8
imageio==2.36.0
inflect==7.4.0
ipdb==0.13.13
jaxtyping==0.2.38
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
legacy-api-wrap==1.4
leidenalg==0.10.2
llvmlite==0.43.0
locket==1.0.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-scalebar==0.8.1
mdurl==0.1.2
mkl-service==2.4.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multipledispatch==1.0.0
multiprocess==0.70.16
multiscale_spatial_image==1.0.1
natsort==8.4.0
networkx==3.4.2
ninja==1.11.1.3
nmslib==2.1.1
numba==0.60.0
numcodecs==0.13.1
numexpr==2.10.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nystrom-attention==0.0.12
ome-zarr==0.9.0
omnipath==1.0.8
opencv-python==4.10.0.84
openslide-python==1.4.1
osqp==0.6.7.post3
pandas==2.2.3
param==2.1.1
partd==1.4.2
patsy==0.5.6
pickleshare==0.7.5
pillow==11.0.0
PIMS==0.7
platformdirs==3.0.0
pooch==1.8.2
POT==0.9.4
propcache==0.2.0
protobuf==5.28.3
psutil==5.9.4
pure-eval==0.2.2
pyarrow==18.0.0
pybind11==2.6.1
pyct==0.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pyg-lib==0.4.0+pt25cpu
Pygments==2.14.0
pynndescent==0.5.13
pyogrio==0.10.0
pyparsing==3.2.0
pyproj==3.7.0
python-dateutil==2.8.2
pytz==2024.2
PyYAML==6.0.2
pyzmq==25.0.0
qdldl==0.1.7.post4
regex==2024.9.11
requests==2.32.3
rich==13.9.4
s3fs==2023.6.0
safetensors==0.4.5
scanpy==1.10.3
scikit-image==0.24.0
scikit-learn==1.5.2
scikit-survival==0.23.0
scipy==1.14.1
seaborn==0.13.2
sentencepiece==0.2.0
sentry-sdk==2.22.0
session_info==1.0.0
setproctitle==1.3.5
shapely==2.0.6
slicerator==1.1.0
smmap==5.0.2
sortedcontainers==2.4.0
sparse_autoencoder==1.10.0
spatial_image==1.1.0
spatialdata==0.2.3
spatialleiden==0.1.1
squidpy==1.6.1
statsmodels==0.14.4
stdlib-list==0.11.0
StrEnum==0.4.15
sympy==1.13.1
tblib==3.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2024.9.20
tokenizers==0.20.1
tomli==2.0.2
toolz==1.0.0
torch==2.5.1+cu124
torch-geometric==2.6.1
torch_cluster==1.6.3+pt25cpu
torch_scatter==2.1.2+pt25cpu
torch_sparse==0.6.18+pt25cpu
torch_spline_conv==1.2.2+pt25cpu
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tornado==6.2
tqdm==4.66.6
transformer-lens==2.15.0
transformers==4.46.1
transformers-stream-generator==0.0.5
triton==3.1.0
typeguard==4.4.0
tzdata==2024.2
umap-learn==0.5.7
urllib3==1.26.20
validators==0.34.0
wadler_lindig==0.1.3
wandb==0.19.8
Werkzeug==3.1.1
wrapt==1.16.0
xarray==2024.7.0
xarray-dataclasses==1.8.0
xarray-datatree==0.0.15
xarray-schema==0.0.3
xarray-spatial==0.4.0
xformers==0.0.28.post3
xxhash==3.5.0
yarl==1.17.1
zarr==2.18.3
zict==3.0.0
zstandard==0.23.0
